begin session-state tests
