+++ src/features/session_replay/aggregate/index.js
@@ -149,7 +149,7 @@ export class Aggregate extends AggregateBase {
       if (fullSample) this.mode = MODE.FULL // full mode has precedence over error mode
       else if (errorSample) this.mode = MODE.ERROR
       // If neither are selected, then don't record (early return)
-      return
+      else return
     }
 
     // FULL mode records AND reports from the beginning, while ERROR mode only records (but does not report).
@@ -175,6 +175,8 @@ export class Aggregate extends AggregateBase {
     this.isFirstChunk = !!session.isNew
 
     session.state.sessionReplay = this.mode
+
+    console.log('state', session.state.sessionReplay)
   }
 
   prepareHarvest (options) {
