+++ tests/specs/session-replay/index.e2e.js
@@ -1,0 +1,3 @@
+import sessionStateTests from './session-state'
+
+sessionStateTests()
